refactor(view/app): wire nested debug_scene module layout
