+++ view/app/src/app_state/debug_scene/mod.rs
@@ -1,0 +1,7 @@
+//! AppState のデバッグ表示ロード（octree/toolpath/svg/nurbs）を扱うモジュール群。
+
+mod camera_fit;
+mod nurbs;
+mod octree;
+mod shapes;
+mod toolpath;
